Wrote Frame buffer, added SRAM inputs and outputs to top level.
We still need to add some sort of state machine. There will eventually be three different coordinates we need to keep track of. 1 The savex,savey, which keep track of what is being written into the fram buffer memory. I believe that corordiante will also be the one that all of the draw functions need.
draw X, Draw Y, the address that is currently being output from the fifo to the actual screen.
3. Read X ReadY, the adress being read out from the fram buffer to the fifo.
2 and 3 are seperate because first the address will be moved from the SRAM to the fifo, then it will be read from the fifo to the actual VGA. So Read X ready Y may go throguh line one, then Draw X Draw Y will fo through line one